requirement.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
scipy>=1.9.0

# Machine Learning
xgboost>=1.7.0
imbalanced-learn>=0.10.0
scikit-optimize>=0.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-plot>=0.3.0
plotly>=5.10.0

# Configuration & Utilities
PyYAML>=6.0
joblib>=1.2.0
tqdm>=4.64.0

# Advanced Features
shap>=0.41.0
eli5>=0.13.0
lime>=0.2.0

# Development
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0